Refactor clique bounds
