+++ src/bounds.rs
@@ -1,10 +1,12 @@
-//! Bounds for identifying assembly states (i.e., collections of fragments)
-//! from which no further assembly index improvements can be found.
+//! TODO
 
 use bit_set::BitSet;
 use clap::ValueEnum;
 
-use crate::molecule::{Bond, Element, Molecule};
+use crate::{
+    molecule::{Bond, Element, Molecule},
+    reductions::{CompatGraph},
+};
 
 /// Bounding strategies for the search phase of assembly index calcluation.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum)]
@@ -39,20 +41,18 @@ struct EdgeType {
 pub fn bound_exceeded(
     mol: &Molecule,
     fragments: &[BitSet],
-    state_index: usize,
-    best_index: usize,
+    ix: usize,
+    best: usize,
     largest_remove: usize,
     bounds: &[Bound],
 ) -> bool {
     for bound_type in bounds {
         let exceeds = match bound_type {
-            Bound::Log => state_index - log_bound(fragments) >= best_index,
-            Bound::Int => state_index - int_bound(fragments, largest_remove) >= best_index,
-            Bound::VecSimple => {
-                state_index - vec_simple_bound(fragments, largest_remove, mol) >= best_index
-            }
+            Bound::Log => ix - log_bound(fragments) >= best,
+            Bound::Int => ix - int_bound(fragments, largest_remove) >= best,
+            Bound::VecSimple => ix - vec_simple_bound(fragments, largest_remove, mol) >= best,
             Bound::VecSmallFrags => {
-                state_index - vec_small_frags_bound(fragments, largest_remove, mol) >= best_index
+                ix - vec_small_frags_bound(fragments, largest_remove, mol) >= best
             }
             _ => {
                 panic!("One of the chosen bounds is not implemented yet!")
@@ -210,4 +210,195 @@ pub fn vec_small_frags_bound(fragments: &[BitSet], m: usize, mol: &Molecule) -> 
 
     s - (z + size_two_types.len() + size_two_fragments.len())
         - ((sl - z) as f32 / m as f32).ceil() as usize
+}
+
+pub fn remaining_weight_bound(graph: &CompatGraph, subgraph: &BitSet) -> usize {
+    let deg_sum = subgraph.iter().map(|v| graph.degree(v, subgraph)).sum::<usize>() as f32;
+    let max_clique = ((1_f32 + (4_f32 * deg_sum + 1_f32).sqrt()) / 2_f32).floor() as usize;
+    let mut sum = 0;
+    let mut iter = subgraph.iter();
+    for _ in 0..max_clique {
+        sum += graph.weight(iter.next().unwrap());
+    };
+
+    sum
+}
+
+pub fn color_bound(graph: &CompatGraph, subgraph: &BitSet) -> usize{
+    // Greedy coloring
+    let mut colors: Vec<i32> = vec![-1; graph.len()];
+    let mut num_colors = 0;
+    let mut largest: Vec<usize> = Vec::new();
+    
+
+    for v in (0..graph.len()).rev() {
+        if !subgraph.contains(v) {
+            continue;
+        }
+
+        let mut used: Vec<usize> = vec![0; num_colors];
+
+        for u in subgraph.intersection(graph.compatible_with(v)) {
+            if colors[u] != -1 {
+                used[colors[u] as usize] = 1;
+            }
+        }
+
+        let mut max = 0;
+        let mut max_idx = num_colors;
+        for i in 0..num_colors {
+            if used[i] == 0 && largest[i] > max {
+                max = largest[i];
+                max_idx = i;
+            }
+        }
+
+        if max_idx == num_colors {
+            num_colors += 1;
+            largest.push(0);
+        }
+        if graph.weight(v) > largest[max_idx] {
+            largest[max_idx] = graph.weights(v)
+        }
+
+        colors[v] = max_idx as i32;
+    }
+
+    largest.iter().sum::<usize>()
+}
+
+pub fn cover_bound(graph: &CompatGraph, subgraph: &BitSet, sort: bool) -> usize {
+    // Sort vertices
+    if sort {
+        let mut vertices: Vec<(usize, usize)> = Vec::with_capacity(subgraph.len());
+        for v in subgraph {
+            vertices.push((v, graph.degree(v, subgraph)));
+        }   
+        vertices.sort_by(|a, b| b.1.cmp(&a.1));
+        cover_bound_helper(graph, subgraph, vertices.iter().map(|(v, _)| *v))
+    }
+    else {
+        let vertices = (0..graph.len()).rev().filter(|v| subgraph.contains(*v));
+        cover_bound_helper(graph, subgraph, vertices)
+    }
+}
+
+fn cover_bound_helper(graph: &CompatGraph, subgraph: &BitSet, iter: impl Iterator<Item = usize>) -> usize {
+    let mut colors: Vec<Option<Vec<usize>>> = vec![None; graph.len()];
+    let mut col_weights = vec![];
+    let mut num_col = 0;
+
+    for v in iter {
+        let mut v_col = Vec::new();
+        let mut used = vec![0; num_col];
+
+        // Find colors used in neighborhood of v
+        for u in subgraph.intersection(graph.compatible_with(v)) {
+            let Some(u_col) = &colors[u] else {
+                continue;
+            };
+
+            for c in u_col {
+                used[*c] = 1;
+            }
+        }
+
+        let mut total_weight = 0;
+        let v_val = graph.weight(v);
+        // Find colors to give to v
+        for c in 0..num_col {
+            if used[c] == 1 {
+                continue;
+            }
+
+            v_col.push(c);
+            total_weight += col_weights[c];
+
+            if total_weight >= v_val {
+                break;
+            }
+        }
+
+        if total_weight == 0 {
+            v_col.push(num_col);
+            col_weights.push(v_val);
+            num_col += 1
+        }
+        else if total_weight < v_val {
+            let mut k = num_col - 1;
+            while used[k] == 1 {
+                k -= 1
+            }
+            col_weights[k] += v_val - total_weight
+        }
+
+        colors[v] = Some(v_col);
+    };
+
+    col_weights.iter().sum()
+}
+
+pub fn frag_bound(graph: &CompatGraph, subgraph: &BitSet, fragments: &[BitSet]) -> usize {
+    let total_bonds = fragments.iter().map(|x| x.len()).sum::<usize>();
+    let mut bound = 0;
+    let sizes = {
+        let mut vec = vec![];
+        let mut prev = 0;
+        for s in subgraph.iter().map(|v| graph.weight(v) + 1) {
+            if s != prev {
+                vec.push(s);
+                prev = s;
+            }
+        }
+
+        vec
+    };
+    
+    for i in sizes {
+        let mut bound_temp = 0;
+        let mut has_bonds = fragments.len();
+        let mut num_bonds: Vec<usize> = fragments.iter().map(|x| x.len()).collect();
+        let mut smallest_remove = i;
+
+        for v in subgraph.iter() {
+            if has_bonds == 0 {
+                break;
+            }
+            if graph.weight(v) + 1 > i {
+                continue;
+            }
+
+
+            let dup = &graph.matches(v);
+            let bond = dup.1.iter().next().unwrap();
+            let mut j = 0;
+            while !fragments[j].contains(bond) {
+                j += 1;
+            }
+
+            if num_bonds[j] > 0 {
+                let remove = std::cmp::min(dup.0.len(), num_bonds[j]);
+                bound_temp += 1;
+                num_bonds[j] -= remove;
+                smallest_remove = std::cmp::min(smallest_remove, remove);
+
+                if num_bonds[j] == 0 {
+                    has_bonds -= 1;
+                }
+            }
+        }
+
+        let leftover = num_bonds.iter().sum::<usize>();
+        let log = {
+            if leftover > 0 {
+                0
+            }
+            else {
+                (smallest_remove as f32).log2().ceil() as usize
+            }
+        };
+        bound = std::cmp::max(bound, total_bonds - bound_temp - leftover - log);
+    }
+
+    bound
 }
